+++ ClientApp/src/components/tests/EDRSRoundtripProducing.js
@@ -6,6 +6,7 @@ import axios from 'axios';
 import { toast } from 'react-semantic-toasts';
 import { FHIRInfo } from '../misc/info/FHIRInfo';
 import { Getter } from '../misc/Getter';
+import report from'../report';
 
 export class EDRSRoundtripProducing extends Component {
   displayName = EDRSRoundtripProducing.name;
@@ -98,6 +99,13 @@ export class EDRSRoundtripProducing extends Component {
     });
   }
 
+  downloadAsFile(contents) {
+    var element = document.createElement('a');
+    element.setAttribute('href', 'data:text/plain;charset=utf-8,' + encodeURIComponent(contents));
+    element.setAttribute('download', `canary-report-${new Date().getTime()}.html`);
+    element.click();
+  }
+
   render() {
     return (
       <React.Fragment>
@@ -135,6 +143,9 @@ export class EDRSRoundtripProducing extends Component {
                     <Statistic.Label>Incorrect</Statistic.Label>
                   </Statistic>
                 </Statistic.Group>
+                <Grid centered columns={1} className="p-t-30 p-b-15">
+                  <Button icon labelPosition='left' primary onClick={() => this.downloadAsFile(report(this.state.test, null))}><Icon name='download' />Generate Downloadable Report</Button>
+                </Grid>
                 <div className="p-b-20" />
                 <Form size="large">
                   <FHIRInfo fhirInfo={this.state.test.results} editable={false} testMode={true} />